Import ACPICA 20140424.

+++ source/compiler/preprocess.h
@@ -257,7 +257,7 @@ PrGetNextToken (
 void
 PrError (
     UINT8                   Level,
-    UINT8                   MessageId,
+    UINT16                  MessageId,
     UINT32                  Column);
 
 void
